Move provider scope into its own file
