+++ shared/extension/behaviour/ProviderScope.hpp
@@ -1,0 +1,115 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2024 NUbots
+ *
+ * This file is part of the NUbots codebase.
+ * See https://github.com/NUbots/NUbots for further info.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+#ifndef EXTENSION_BEHAVIOUR_PROVIDER_SCOPE_HPP
+#define EXTENSION_BEHAVIOUR_PROVIDER_SCOPE_HPP
+
+#include <nuclear>
+
+#include "RunReason.hpp"
+#include "commands.hpp"
+
+namespace extension::behaviour {
+    /**
+     * This struct holds all the scope information for a provider.
+     *
+     * While running in a ProviderScope, the behaviour of task emits will be different.
+     */
+    struct ProviderScope {
+        ProviderScope(const std::type_index& provider_type, NUClear::threading::ReactionTask& reaction_task)
+            : provider_type(provider_type)
+            , reaction_task(reaction_task)
+            , previous_scope(std::exchange(current_scope, this))
+            , tasks(std::make_unique<std::vector<commands::BehaviourTask>>()) {}
+
+        ProviderScope(ProviderScope&& other)
+            : provider_type(other.provider_type)
+            , reaction_task(other.reaction_task)
+            , previous_scope(std::exchange(other.previous_scope, nullptr))
+            , tasks(std::move(other.tasks)) {}
+
+        ProviderScope& operator=(ProviderScope&& other) = delete;
+        ProviderScope(const ProviderScope&)             = delete;
+        ProviderScope& operator=(const ProviderScope&)  = delete;
+
+        ~ProviderScope() {
+            // Upon desctruction restore the previous scope as the current scope
+            current_scope = previous_scope;
+
+            // Emit all the tasks that were accumulated in this scope
+            reaction_task.parent->reactor.emit(std::make_unique<commands::BehaviourTasks>(provider_type,
+                                                                                          reaction_task.parent->id,
+                                                                                          reaction_task.id,
+                                                                                          false,
+                                                                                          std::move(*tasks)));
+        }
+
+        template <typename T>
+        static void emit(NUClear::PowerPlant& powerplant,
+                         const std::shared_ptr<T>& data,
+                         const int priority,
+                         const bool optional,
+                         const std::string& name) {
+
+            // Create the task object
+            auto task = commands::BehaviourTask(typeid(T), data, name, priority, optional);
+
+            // If we are in a scope (running in a provider) accumulate the task to be emitted later
+            if (current_scope != nullptr) {
+                current_scope->tasks->push_back(std::move(task));
+            }
+            /// Root tasks emit the task immediately as a single pack
+            else {
+                powerplant.emit(std::make_unique<commands::BehaviourTasks>(typeid(commands::RootProvider<T>),
+                                                                           0,
+                                                                           0,
+                                                                           true,
+                                                                           std::vector<commands::BehaviourTask>{task}));
+            }
+        }
+
+    private:
+        /// The type of the provider that this scope is for
+        std::type_index provider_type;
+        /// The reaction task object this scope is for
+        NUClear::threading::ReactionTask& reaction_task;
+        /// The previous scope that was active before this one so it can be restored
+        /// This should only not be nullptr if somehow an Inline emit forces another provider to run while still in
+        /// another provider
+        ProviderScope* previous_scope = nullptr;
+        /// The tasks that have been accumulated in this scope
+        std::unique_ptr<std::vector<commands::BehaviourTask>> tasks{};
+
+        /// Holds the scope object which is currently active
+        static thread_local ProviderScope* current_scope;
+    };
+
+    // Define and initialize the static thread-local variable outside the class
+    thread_local ProviderScope* ProviderScope::current_scope = nullptr;
+
+}  // namespace extension::behaviour
+
+#endif  // EXTENSION_BEHAVIOUR_PROVIDER_SCOPE_HPP
